implement auth usin police

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,10 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
 {
+
+    public function __construct(){
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -67,7 +71,7 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        
+        $this->authorize('update',$question);
         return view('questions.edit',compact('question'));
     }
 
@@ -80,6 +84,7 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        $this->authorize('update',$question);
         $question->update($request->only('title','body'));
         return redirect()->route('questions.index')->with('success',"Your question has been Updated");
     }
@@ -91,9 +96,9 @@ class QuestionsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Question $question)
-    {
+    {   
+        $this->authorize('delete',$question);
         $question->delete();
-
         return redirect()->route('questions.index')->with('success',"Your question has been Deleted");
     }
 }
